requirements.txt
streamlit>=1.23.0
python-dotenv>=1.0.0
requests>=2.25.0
